# generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Shelfmark API",
    "version": "0.1.0",
    "description": "HTTP API for the running Shelfmark instance. Generated from Flask routes. There is no separate published spec. GET /openapi.json and GET /api/openapi.json are unauthenticated. Most other /api/* routes use login_required when authentication is enabled. ISBN metadata search is GET /api/metadata/search?query=; the handler does not accept search_type. ISBN file search is GET /api/releases with isbn= and source=direct_download."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/admin/download-defaults": {
      "get": {
        "operationId": "admin_download_defaults_get",
        "summary": "Return default download destination settings for new users.",
        "description": "Return default download destination settings for new users.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/admin/booklore-options": {
      "get": {
        "operationId": "admin_booklore_options_get",
        "summary": "List Booklore libraries for admin settings.",
        "description": "List Booklore libraries for admin settings.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/admin/users/{user_id}/search-preferences": {
      "get": {
        "operationId": "admin_get_search_preferences_get",
        "summary": "Return search preferences for one user.",
        "description": "Return search preferences for one user.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/admin/users/{user_id}/notification-preferences": {
      "get": {
        "operationId": "admin_get_notification_preferences_get",
        "summary": "Return notification preferences for one user.",
        "description": "Return notification preferences for one user.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/admin/users/{user_id}/notification-preferences/test": {
      "post": {
        "operationId": "admin_test_notification_preferences_post",
        "summary": "Send a test notification for one user.",
        "description": "Send a test notification for one user.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/admin/settings/overrides-summary": {
      "get": {
        "operationId": "admin_settings_overrides_summary_get",
        "summary": "Summarize per-user settings overrides.",
        "description": "Summarize per-user settings overrides.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/admin/users/{user_id}/effective-settings": {
      "get": {
        "operationId": "admin_get_effective_settings_get",
        "summary": "Return merged settings for one user.",
        "description": "Return merged settings for one user.",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/api/settings/{tab_name}": {
      "get": {
        "operationId": "api_settings_get_tab_get",
        "summary": "Get settings for a specific tab.",
        "description": "Get settings for a specific tab.\n\nPath Parameters:\n    tab_name (str): Settings tab name (e.g., \"general\", \"hardcover\")\n\nReturns:\n    flask.Response: JSON with tab settings and values.",
        "parameters": [
          {
            "name": "tab_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "operationId": "api_settings_update_tab_put",
        "summary": "Update settings for a specific tab.",
        "description": "Update settings for a specific tab.\n\nPath Parameters:\n    tab_name (str): Settings tab name\n\nRequest Body:\n    JSON object with setting keys and values to update.\n\nReturns:\n    flask.Response: JSON with update result.",
        "parameters": [
          {
            "name": "tab_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Authentication required when AUTH_METHOD is not none"
          },
          "404": {
            "description": "Not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  }
}